created spots seeder data

+++ backend/db/models/spot.js
@@ -38,10 +38,10 @@ module.exports = (sequelize, DataTypes) => {
         allowNull: false
       },
       lat: {
-        type: DataTypes.INTEGER
+        type: DataTypes.DECIMAL
       },
       lng: {
-        type: DataTypes.INTEGER
+        type: DataTypes.DECIMAL
       },
       name: {
         type: DataTypes.STRING,
